tests: enable building root filesystem with dracut
So far, install_kata_image.sh only supported building the root filesystem
image or initrd using the "distro" method.  This method however depends on
being able to run a container which causes trouble if the building process
runs in a container already as is the case for Openshift CI.

The dracut-based build method doesn't need to run a container and is thus
better suited for such an environment.

An input variable IGNORE_CACHED_ARTIFACTS for install_kata_image.sh was also
added which can be used to force local root filesystem build (as opposed using
a cached artifact).  This is mostly useful for testing.

Signed-off-by: Pavel Mores <pmores@redhat.com>