tests: enable building root filesystem with dracut
So far, install_kata_image.sh only supported building the root filesystem
image or initrd using the "distro" method.  This method however depends on
being able to run a container which causes trouble if the building process
runs in a container already as is the case for Openshift CI.

The dracut-based build method doesn't need to run a container and is thus
better suited for such an environment.

Signed-off-by: Pavel Mores <pmores@redhat.com>