integration/stability: Ignore running hypervisor processes
Currently the statility test script is going to fail if there is any hypervisor
process running on the system, but this situation can happen quit often, for
instance, in case of leftovers from previously executed tests.

Instead of failing the test, let's ignore the PIDs of any already running hypervisor
process.

Fixes #3654
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>